[Fix] map-generation: the generated maps ship their own tilesets
Copying dest/map/main/official/ into a datapack as map/main/generated/ gave
"tileset not found": every map referenced ../../../tileset/x.tsx, i.e. the
datapack-wide map/tileset/. That dir is READ-ONLY for the generator and owns
none of the 25 sheets the tool uses (gym*, heal*, shop*, houses*, building*,
desert*, indoor, mediteranian, sea), so the maps loaded with no tileset at all.
Only the previous run directory made it work, because the tool's own tileset/
happened to sit next to the binary.

dest/map/main/official/ is ONE datapack map label, copied under map/main/<label>/,
so it now carries its own tileset/ and every reference is ../tileset/x.tsx: the
folder is self-contained, droppable into any datapack, and map/tileset/ is never
written. dest/map/tileset/ and dest/map/main/tileset/ stay the run staging pool;
LoadMap::shipTileset copies out of it at WRITE time, so a tileset no map
references is never shipped (35 shipped, not the 43 collected).

Shipping at READ time instead was tried first and hung the generator: a template
tileset left on its source path stops matching the world one in
MapBrush::tiledMapToMapTemplate, which re-adds it per template until the world
passes 255 tilesets, where the uint8_t loop counter wraps and spins for ever
(12 min, 18 GB). Those two counters are int now, so that class of mistake ends
instead of eating the machine.

check-generated.py refuses a reference leaving the label - the guard that was
missing, it flags the broken copy on the first map. install-generated.py is the
reviewable install: it refuses to touch the datapack while a reference escapes,
lists what it replaces, and skips the all.tmx debug dump.

Verified: 617 maps, 652 teleports, 0 problems; 11513 tileset references, 0
broken; renders correct from inside the datapack; fresh build dir generates in
one command; repeated runs byte identical (10.4 s).

+++ tools/map-procedural-generation-terrain/MapBrush.cpp
@@ -157,7 +157,11 @@ MapBrush::MapTemplate MapBrush::tiledMapToMapTemplate(Tiled::Map *templateMap,Ti
         }
     }
     {
-        uint8_t templateTilesetIndex=0;
+        //int, NOT uint8_t: tilesetCount() is an int, and a 8 bit counter wraps at
+        //255 into an ENDLESS loop instead of ending. A tileset pairing that stops
+        //matching pushes the world past 255 tilesets, and the generator then spins
+        //for ever eating memory instead of reporting anything.
+        int templateTilesetIndex=0;
         while(templateTilesetIndex<templateMap->tilesetCount())
         {
             Tiled::SharedTileset tileset=templateMap->tilesetAt(templateTilesetIndex);
@@ -166,7 +170,7 @@ MapBrush::MapTemplate MapBrush::tiledMapToMapTemplate(Tiled::Map *templateMap,Ti
 
             QFileInfo tilesetFile(tileset->fileName());
             QString tilesetPath=tilesetFile.absoluteFilePath();
-            uint8_t templateTilesetWorldIndex=0;
+            int templateTilesetWorldIndex=0;
             //search into current tileset
             while(templateTilesetWorldIndex<worldMap.tilesetCount())
             {
